merge

+++ include/Utils.hpp
@@ -8,6 +8,7 @@
 # include <utility>
 # include <stdlib.h>
 # include <sys/stat.h>
+# include <sys/time.h>
 # include <ctime>
 # include "Status.hpp"
 # include "Enum.hpp"
@@ -42,9 +43,9 @@ namespace utils
 	std::string					DivideStrByCRLF(std::string &data);
 	std::string					DivideStrBySpace(std::string &data);
 	void						TrimSpaceTap(std::string &trgt);
-	int							hstoi(const std::string &trgt);
-	int							ReadChunkSize(std::string &data);
-	std::string					ReadData(std::string &data, int size);
+	size_t						hstoi(const std::string &trgt);
+	size_t						ReadChunkSize(std::string &data);
+	std::string					ReadData(std::string &data, size_t size);
 	bool						CheckNameChar(std::string &data);
 	bool						CheckHostDup(std::string &header_name, std::map<std::string, std::string> &request_header);
 	void						SplitHeaderData(std::string &data, std::string &name, std::string &value);
